+++ app/src/main/java/com/forrestguice/suntimeswidget/alarmclock/bedtime/BedtimeViewHolder.java
@@ -16,7 +16,7 @@
     along with SuntimesWidget.  If not, see <http://www.gnu.org/licenses/>.
 */ 
 
-package com.forrestguice.suntimeswidget.alarmclock.ui.bedtime;
+package com.forrestguice.suntimeswidget.alarmclock.bedtime;
 
 import android.annotation.SuppressLint;
 import android.content.Context;
