+++ app/src/main/java/com/forrestguice/suntimeswidget/settings/WidgetActions.java
@@ -34,8 +34,7 @@ import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.forrestguice.suntimeswidget.alarmclock.AlarmNotifications;
-import com.forrestguice.suntimeswidget.alarmclock.ui.bedtime.BedtimeActivity;
-import com.forrestguice.suntimeswidget.alarmclock.ui.bedtime.BedtimeBroadcastReceiver;
+import com.forrestguice.suntimeswidget.alarmclock.bedtime.BedtimeBroadcastReceiver;
 import com.forrestguice.suntimeswidget.calculator.SuntimesClockData;
 import com.forrestguice.suntimeswidget.calculator.SuntimesEquinoxSolsticeData;
 import com.forrestguice.suntimeswidget.calculator.SuntimesMoonData;
