+++ app/src/main/res/layout/layout_activity_bedtime.xml
@@ -49,7 +49,7 @@
 			app:layout_constraintTop_toBottomOf="@+id/app_menubar"
 			android:paddingTop="4dp">
 
-			<fragment android:name="com.forrestguice.suntimeswidget.alarmclock.ui.bedtime.BedtimeDialog"
+			<fragment android:name="com.forrestguice.suntimeswidget.alarmclock.bedtime.BedtimeDialog"
 				android:id="@+id/listFragment"
 				android:layout_width="match_parent" android:layout_height="match_parent" />
